cli/requirements.txt
typer>=0.9.0
rich>=13.7.0
requests>=2.31.0
openai>=1.14.0
mcp>=1.0.0
python-dotenv>=1.0.1
